# openapi.yaml
openapi: 3.0.3
info:
  title: gdl-api
  description: gdl-api express api server + frontend + bots
  version: 0.2.31
  license:
    name: MIT
servers:
  - url: https://gdl.mcalec.dev/api
    description: production server
  - url: http://localhost:3030/api
    description: local server
...
paths:
...
  /files/:
    get:
      tags:
        - Files
      summary: Get all files and directories from the root
      description: List all files and directories from the BASE_DIR with optional sorting and pagination
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, size, date, type]
          description: Field to sort by
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction
        - in: query
          name: limit
          schema:
            type: integer
          description: Maximum number of results (up to PAGINATION_LIMIT)
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
      responses:
        "200":
          description: List of files and directories
        "403":
          description: User not authenticated
        "500":
          description: Internal server error
  /files/{collection}/:
    get:
      tags:
        - Files
      summary: Get files from a specific collection
      description: List files and subdirectories within a specific collection
      parameters:
        - in: path
          name: collection
          required: true
          schema:
            type: string
          description: Collection name
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, size, date, type]
          description: Field to sort by
        - in: query
          name: direction
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction
      responses:
        "200":
          description: List of files in collection
        "400":
          description: Invalid path parameters
        "404":
          description: Collection not found
  /files/{collection}/{author}/:
    get:
      tags:
        - Files
      summary: Get files from a specific author within a collection
      description: List files from a specific author directory within a collection
      parameters:
        - in: path
          name: collection
          required: true
          schema:
            type: string
          description: Collection name
        - in: path
          name: author
          required: true
          schema:
            type: string
          description: Author/subdirectory name
      responses:
        "200":
          description: List of files from author directory
        "404":
          description: Path not found
  /download/:
    get:
      tags:
        - Downloads
      summary: Download a file using its UUID
      description: Downloads a file from the server using the provided UUID. Supports both GET and POST methods.
      parameters:
        - in: query
          name: uuid
          required: true
          schema:
            type: string
          description: The UUID of the file to download
      responses:
        "200":
          description: File content as binary data
        "400":
          description: Missing or invalid uuid parameter
        "404":
          description: File not found or not accessible
        "500":
          description: Internal server error
    post:
      tags:
        - Downloads
      summary: Download a file using its UUID (POST method)
      description: Downloads a file from the server using the provided UUID via POST body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: The UUID of the file to download
              required:
                - uuid
      responses:
        "200":
          description: File content as binary data
        "400":
          description: Missing or invalid uuid parameter
        "404":
          description: File not found or not accessible
        "500":
          description: Internal server error
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authenticated requests
security:
  - cookieAuth: []